full work

+++ ft_printf.h
@@ -6,14 +6,13 @@
 /*   By: sabdulba <sabdulba@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/08 19:45:20 by sabdulba          #+#    #+#             */
-/*   Updated: 2024/11/09 16:56:38 by sabdulba         ###   ########.fr       */
+/*   Updated: 2024/11/09 22:19:48 by sabdulba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
-//remove stdio before push
-# include <stdio.h>
+
 # include <unistd.h>
 # include <stdarg.h>
 
@@ -22,6 +21,6 @@ void	ft_putchar(char c, int *count);
 void	ft_putstr(char *str, int *count);
 void	ft_putnbr(int n, int *count);
 void	ft_unsigned(unsigned int n, int *count);
-void	ft_hex(unsigned long n, int *count, char up_low);
+void	ft_hex(unsigned int n, int *count, char up_low);
 void	ft_pointer(size_t point, int *count);
 #endif
